Move NetworkContext to context/network
