+++ src/components/BridgeForm.js
@@ -30,7 +30,7 @@ import {
     TextInput
 } from '@patternfly/react-core';
 import cockpit from 'cockpit';
-import { useNetworkDispatch, useNetworkState, addConnection, updateConnection } from '../NetworkContext';
+import { useNetworkDispatch, useNetworkState, addConnection, updateConnection } from '../context/network';
 import interfaceType from '../lib/model/interfaceType';
 
 const _ = cockpit.gettext;
